Move query into Modal file because it's cleaner
