+++ src/components/PatchActionButtons/EnqueuePatch.tsx
@@ -1,15 +1,6 @@
 import { useState } from "react";
-import { useQuery } from "@apollo/client";
 import { DropdownItem } from "components/ButtonDropdown";
-import {
-  CodeChangesQuery,
-  CodeChangesQueryVariables,
-} from "gql/generated/types";
-import { GET_CODE_CHANGES } from "gql/queries";
 import { EnqueuePatchModal } from "pages/version/index";
-import { commits } from "utils";
-
-const { shouldPreserveCommits } = commits;
 
 interface EnqueuePatchProps {
   patchId: string;
@@ -32,19 +23,6 @@ export const EnqueuePatch: React.VFC<EnqueuePatchProps> = ({
       ? visibilityControl
       : fallbackVisibilityControl;
 
-  const { data, previousData } = useQuery<
-    CodeChangesQuery,
-    CodeChangesQueryVariables
-  >(GET_CODE_CHANGES, {
-    variables: { id: patchId },
-    skip: !isVisible,
-  });
-  const { patch } = data ?? previousData ?? {};
-  const { moduleCodeChanges } = patch ?? {};
-  const preserveCommits = moduleCodeChanges?.length
-    ? shouldPreserveCommits(moduleCodeChanges[0].fileDiffs)
-    : false;
-
   return (
     <>
       <DropdownItem
@@ -60,7 +38,6 @@ export const EnqueuePatch: React.VFC<EnqueuePatchProps> = ({
         visible={isVisible}
         onFinished={() => setIsVisible(false)}
         refetchQueries={refetchQueries}
-        preserveCommits={preserveCommits}
       />
     </>
   );
